I'm doing this my first time ever i really did git commit, not really sure what to put here. But yea, this is yet another sequencer. I didn't come with a better name yet. It's still a work in progress, taking much longer than I initially expected. It currently features a Arduino MEGA as it has multiple hardware serial ports so I can read multiple midi inputs. One only listens to time and transport messages, the other one is sequenced. I use it to sequence my MicoKORG. Oh, there's also a rotary encoder that I wanted to use for overall control, even writing notes, but these features now don't get much attention if any.
I want this to be the most idiot-proof thing ever. Push a record button, input sequence, play it. If the clock is ticking and a midi start byte has been received. If you long-press the rotary encoder, the device goes to edit mode where you scroll steps and add and delete notes. If you push the button shortly again you insert or edit currently recorded note.

Getting to the best part is the recording mechanism that is - to my taste - not very sophisticated, I'm sorry, I'm not a very good programmer.

Now lot's of things are going to change as my Teensy 3.6, Adafruit NeoTrellis and a couple of RGB rotary encoders arrive. So it's mainly the recording mechanism that I'm working on currently. I want to add quantisation later but it's currently not on the top of the list. See, MEGA does not have enough RAM to hold the 4 arrays of 256 notes which I wanted to do first. I have a perfectly working 2 arrays but even 3 is too much. Now after the edits I made (like adding queues, love those, I'll probably get to them) I can see that the note structure could lack a couple of variables which would be nice. Also storing the millis_started thing to something smaller is an idea. So optimise as much as possible to have a fast and efficient code so I can do some messy stuff with the Teensy, that thing is a beast!

No, I do not have a girlfriend.

I mean... Who's gonna read this?

I only display BPM on the 16x2 char display currently as I felt the time spent on printing was too much to waste. Now it only displays the BPM when it changes so I have time for other stuff. But I'm ordering also a 14 segment 4 char display from Adafruit so that's gonna be big. Ever heard of Novation Circuit? I'm basically inventing it again for pure sequencing. I plan on getting the Behringer Neutron thingy so I want the box to sequence it as well. With the MicroKORG keyboard, of course. I feel inspiration in the ease of use, that's something Novation does well with those machines and I want to kinda be the same.

I also want to sequence automation, naturally. LOTS OF AUTOMATION!

I started working on this idea of recording the sequences in a sequence, polyphonically. Polyphonically? Well. Just that I would never fill a 4x256x12B array! What a waste of memory, right. RIGHT?! So it's in this big array. The notes are stored sequentially, sorted by beat. One beat is a sixteenth note. I use quite mismatched terms sometimes, sorry, I'm not a native speaker. And I'm high as fuck.

My target is big. I want people to want it. Not for profit (though that would be nice) but simply for it prooving that I made it really usable. And perhaps play a little with the DSP stuff in the teensy. If I finish this I'm gonna be proud of myself.

So much to this yet another sequencer.